cnn_model/requirements.txt
tensorflow
keras
gradio
pillow
numpy
